+++ include/ignition/rendering/RenderPassSystem.hh
@@ -22,6 +22,8 @@
 #include <string>
 #include <typeinfo>
 
+#include <ignition/common/SuppressWarning.hh>
+
 #include "ignition/rendering/config.hh"
 #include "ignition/rendering/Export.hh"
 #include "ignition/rendering/RenderPass.hh"
@@ -77,9 +79,14 @@ namespace ignition
       /// \return Pointer to the render pass created
       private: RenderPassPtr CreateImpl(const std::string &_type);
 
+      /// \brief A map of render pass type id name to its factory class
+      IGN_COMMON_WARN_IGNORE__DLL_INTERFACE_MISSING
+      private: static std::map<std::string, RenderPassFactory *> renderPassMap;
+
       /// \internal
       /// \brief Pointer to private data class
       private: std::unique_ptr<RenderPassSystemPrivate> dataPtr;
+      IGN_COMMON_WARN_RESUME__DLL_INTERFACE_MISSING
     };
 
     }